[Web] Feat : :zap:️ add validation to socket #83
소켓 통신에서 받는 데이터에 대한 유효성 검사 추가
error event에 대한 예외 처리 추가

+++ server/src/sockets/socketController.ts
@@ -4,39 +4,72 @@ import socketService from './socketService';
 import socketRoomModel from '../models/socketRoomModel';
 import roomSocketsInfoModel from '../models/roomSocketsInfoModel';
 import roomInfoModel from '../models/roomInfoModel';
-import { UserDataType, SendChatType } from '../@types/socketType';
+import validationUtil from '../utils/validation';
+import SocketErrorMessage from '../@types/socketErrorMessage';
+import { UserDataType, SendChatType } from '../@types/dataType';
 
 const enterChatroom = async (socket: Socket, io: SocketIO.Server, userData: UserDataType) => {
   const { roomCode, nickname, language } = userData;
+
+  if (validationUtil.isRoomCodeValid(roomCode))
+    return socketService.emitSocketError(socket, SocketErrorMessage.ROOM_CODE);
+  if (validationUtil.isNicknameValid(nickname))
+    return socketService.emitSocketError(socket, SocketErrorMessage.NICKNAME);
+  if (validationUtil.isLanguageValid(language))
+    return socketService.emitSocketError(socket, SocketErrorMessage.LANGUAGE);
+
   socket.join(roomCode);
-  await socketService.insertSocketInfoIntoDB(socket.id, roomCode, nickname, language);
-  const participantsList = await socketService.getParticipantsListFromRoomCode(roomCode, 'enter');
-  io.to(roomCode).emit('receive participants list', participantsList);
+
+  try {
+    await socketService.insertSocketInfoIntoDB(socket.id, roomCode, nickname, language);
+    const participantsList = await socketService.getParticipantsListFromRoomCode(roomCode, 'enter');
+    io.to(roomCode).emit('receive participants list', participantsList);
+  } catch (err) {
+    return socketService.emitSocketError(socket, SocketErrorMessage.SERVER);
+  }
 };
 
 const sendChat = async (socket: Socket, io: SocketIO.Server, sendChat: SendChatType) => {
   const { Korean, English } = sendChat;
-  const { roomCode, receiveChat } = await socketService.createReceiveChatType(socket.id, Korean, English);
-  io.to(roomCode).emit('receive chat', receiveChat);
+
+  if (validationUtil.isMessageValid(sendChat)) return socketService.emitSocketError(socket, SocketErrorMessage.CHAT);
+
+  try {
+    const { roomCode, receiveChat } = await socketService.createReceiveChatType(socket.id, Korean, English);
+    io.to(roomCode).emit('receive chat', receiveChat);
+  } catch (err) {
+    return socketService.emitSocketError(socket, SocketErrorMessage.SERVER);
+  }
 };
 
 const disconnect = async (socket: Socket, io: SocketIO.Server) => {
-  const roomCode: string = await socketRoomModel.getRoomBySocket(socket.id);
-  await socketService.removeSocketInfoFromDB(socket.id, roomCode);
+  try {
+    const roomCode: string = await socketRoomModel.getRoomBySocket(socket.id);
+    if (roomCode !== null) {
+      await socketService.removeSocketInfoFromDB(socket.id, roomCode);
 
-  if (await roomSocketsInfoModel.isRoomEmpty(roomCode)) {
-    await roomInfoModel.removeRoom(roomCode);
-    return;
+      if (await roomSocketsInfoModel.isRoomEmpty(roomCode)) {
+        await roomInfoModel.removeRoom(roomCode);
+        return;
+      }
+
+      const participantsList = await socketService.getParticipantsListFromRoomCode(roomCode, 'leave');
+      io.to(roomCode).emit('receive participants list', participantsList);
+    }
+  } catch (err) {
+    throw new Error(err);
   }
+};
 
-  const participantsList = await socketService.getParticipantsListFromRoomCode(roomCode, 'leave');
-  io.to(roomCode).emit('receive participants list', participantsList);
+const socketError = async (socket: Socket, io: SocketIO.Server) => {
+  await disconnect(socket, io);
 };
 
 const socketController = {
   enterChatroom,
   sendChat,
   disconnect,
+  socketError,
 };
 
 export default socketController;
